requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic-settings==2.2.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
aio-pika==9.4.0
alembic==1.12.1
python-jose==3.3.0
